We can embrace the C language now!
The binary could contain the .ds section. The failure is due to the
size of the binary. There is a big gap between the code and read only
data in the binary. Hence the size of the binary is over than 2Kb now.
Our bootloader assume the size of the binary is less than 2Kb. I also
found although the sector size of the CD is 2Kb, the BIOS read 512b
each time. After fix all of these, we can embrace the C language now!