Fixed the elf/binary conversion problem.
When the initialization value of ds changed from non-zero to zero
the value will move to bss section. The bss section content won't
included in the elf, the binary won't contain it either. It is the
loader's job to prepare the bss area for the executive. Our loader
can't handle this job at this moment, so we let objcopy to take
care of it.